More Testing Programs
Most things are still very early, and so documenting them
very well before I'm set on anything feels like a waste of time.

When things are a bit more "set-in-stone" (which may never happen to be
honest with this hobby project) then I'll actually write down what the
behaviours of some of these programs should be. Most of them are
pretty self explanatory for someone familiar with Unix
(mount and sh diverge greatly from/are way less featured than the
standard Unix/Linux versions)